refactor: reorganize layout folder into functional subfolders
Organized layout components by function for better maintainability:
- header/ - Header navigation components
- footer/ - Footer component
- sidebar/ - Sidebar layout components
- tabs/ - Tab navigation component

File moves:
- layout/Header.tsx → layout/header/Header.tsx
- layout/HeaderDropdown.tsx → layout/header/HeaderDropdown.tsx
- layout/Navigator.tsx → layout/header/Navigator.tsx
- layout/Footer.tsx → layout/footer/Footer.tsx
- layout/SidebarLayout.tsx → layout/sidebar/SidebarLayout.tsx
- layout/SidebarContentLayout.tsx → layout/sidebar/SidebarContentLayout.tsx
- layout/Tabs.tsx → layout/tabs/Tabs.tsx

Updated imports in:
- src/app/layout.tsx (Header, Navigator, Footer)
- src/app/components/sidebar/ProjectSidebar.tsx (SidebarLayout)
- src/app/components/sidebar/TeamSidebar.tsx (SidebarLayout)
- src/app/(box-layout)/team/Team.tsx (SidebarContentLayout)
- src/app/components/project/ProjectPage.tsx (SidebarContentLayout)
- src/app/(box-layout)/portfolio/Portfolio.tsx (Tabs)
- src/app/components/collect/CollectClient.tsx (Tabs)
- src/app/components/collect/constants.ts (TabMode type)

Fixed internal imports to use absolute paths:
- Header.tsx, HeaderDropdown.tsx, Footer.tsx

Verified with pnpm lint and pnpm build.
