+++ src/app/(box-layout)/team/Team.tsx
@@ -5,7 +5,7 @@ import { notFound } from 'next/navigation';
 import TeamSidebar from '@/app/components/sidebar/TeamSidebar';
 import ProfileIcon from '@/app/components/card/portfolio/components/ProfileIcon';
 import ProjectGrid from '@/app/components/project/components/ProjectGrid';
-import SidebarContentLayout from '@/app/components/layout/SidebarContentLayout';
+import SidebarContentLayout from '@/app/components/layout/sidebar/SidebarContentLayout';
 
 const Team = async ({ teamName }: { teamName: string }) => {
   const teamDetail = (await getTeamData(teamName)) ?? notFound();
