+++ src/app/components/project/ProjectPage.tsx
@@ -3,7 +3,7 @@ import ProjectMainContent from '@/app/components/project/components/ProjectMainC
 import ProjectSidebar from '@/app/components/sidebar/ProjectSidebar';
 import { getProjectDetailViewModel } from '@/services/project/getProjectDetail.server';
 import checkProjectEditPermission from '@/services/project/checkProjectEditPermission.server';
-import SidebarContentLayout from '@/app/components/layout/SidebarContentLayout';
+import SidebarContentLayout from '@/app/components/layout/sidebar/SidebarContentLayout';
 import ProfileIcon from '@/app/components/card/portfolio/components/ProfileIcon';
 
 export interface ProjectDetailPageProps {
